fix: fix multiposts and new apscheduler options (#15)
* fix: fix multiposts and new apscheduler options

The theory: I am seeing disconnects in the errbot log that the bot
"restarts" itself. This is "activating" multiple apschedulers because we
are using the in-memory store. This is causing the multiple posts.
Probably was causing the multiple posts with scheduler too.

This change does a couple of things:

* Sets up a webhook that we can use as a bit of a "rpc". The blocker for
using the apscheduler storage backends was trying to pickle the "plugin"
object as a part of the schedule job. This allows us to call a method
that is outside of the class, that can then "call" our class (via the
webhook request).
* Allows configuring the apscheduler by passing in a path to a json
config file. This json file gets read into a python Dict to config like
Method #2 from
https://apscheduler.readthedocs.io/en/stable/userguide.html#configuring-the-scheduler
* Changes config variable names a bit to "namespace" our env variables a
bit
* Adds the ability to "reset" a post so that it can be reposted

* fix: shutdown scheduler on deactivate